backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.8.2
psycopg2-binary==2.9.9
pymupdf==1.24.4
chromadb==0.5.5
openai==1.40.1
google-generativeai==0.7.2
requests==2.32.3
httpx==0.27.0
